+++ ci4/app/Libraries/DeploymentSteps/MigrationJobStep.php
@@ -142,13 +142,16 @@ class MigrationJobStep extends BaseDeploymentStep {
             return 'Database service no longer exists';
         }
 
-        if (!$deployment->domain_id) {
-            return 'Missing domain';
-        }
-        $domain = new Domain();
-        $domain->find($deployment->domain_id);
-        if (!$domain->exists()) {
-            return 'domain no longer exists';
+        $spec = $deployment->findDeploymentSpecification();
+        if ($spec->enable_ingress) {
+            if (!$deployment->domain_id) {
+                return 'Missing domain';
+            }
+            $domain = new Domain();
+            $domain->find($deployment->domain_id);
+            if (!$domain->exists()) {
+                return 'domain no longer exists';
+            }
         }
 
         $namespaceStep = new NamespaceStep();